build: generate sitemap.xml during build
