+++ src/router/prerender.mjs
@@ -1,8 +1,9 @@
-import { mkdir, readdir, readFile, writeFile } from "node:fs/promises";
+import { mkdir, readFile, writeFile } from "node:fs/promises";
 import path from "node:path";
 
-import * as ts from "typescript";
 import { loadEnv } from "vite";
+
+import { collectPageConfigs } from "./pageMeta.mjs";
 
 /* -----------------------------------------------
  * yarn build:prerender コマンドで実行されるスクリプト
@@ -15,74 +16,6 @@ const viteEnv = loadEnv(mode, process.cwd(), "VITE_APP_");
 const SITE_NAME = viteEnv.VITE_APP_SITE_NAME ?? "";
 const SITE_URL = viteEnv.VITE_APP_BASE_URL ?? "";
 const DEFAULT_OG_IMAGE = `${SITE_URL}${viteEnv.VITE_APP_DEFAULT_OG_IMAGE ?? ""}`;
-
-const findPageModules = async (dirPath) => {
-  const entries = await readdir(dirPath, { withFileTypes: true });
-  const pageModules = [];
-
-  for (const entry of entries) {
-    const fullPath = path.join(dirPath, entry.name);
-
-    if (entry.isDirectory()) {
-      pageModules.push(...(await findPageModules(fullPath)));
-      continue;
-    }
-
-    if (entry.isFile() && entry.name === "page.tsx") {
-      pageModules.push(path.relative(process.cwd(), fullPath));
-    }
-  }
-
-  return pageModules;
-};
-
-const findPageMetaNode = (sourceFile) => {
-  for (const statement of sourceFile.statements) {
-    if (!ts.isVariableStatement(statement)) continue;
-
-    const hasExport = statement.modifiers?.some(
-      (modifier) => modifier.kind === ts.SyntaxKind.ExportKeyword,
-    );
-    if (!hasExport) continue;
-
-    for (const declaration of statement.declarationList.declarations) {
-      if (!ts.isIdentifier(declaration.name) || declaration.name.text !== "pageMeta") continue;
-      if (!declaration.initializer) continue;
-      return declaration.initializer;
-    }
-  }
-
-  return null;
-};
-
-const extractPageConfig = async (pagePath) => {
-  const source = await readFile(path.resolve(process.cwd(), pagePath), "utf8");
-  const sourceFile = ts.createSourceFile(
-    pagePath,
-    source,
-    ts.ScriptTarget.Latest,
-    true,
-    ts.ScriptKind.TSX,
-  );
-  const pageMetaNode = findPageMetaNode(sourceFile);
-
-  if (!pageMetaNode) {
-    throw new Error(`pageMeta export not found in ${pagePath}`);
-  }
-
-  const pageMetaSource = sourceFile.text.slice(pageMetaNode.pos, pageMetaNode.end).trim();
-  const pageMeta = new Function(`"use strict"; return (${pageMetaSource});`)();
-
-  if (!pageMeta) {
-    throw new Error(`pageMeta export not found in ${pagePath}`);
-  }
-
-  if (!pageMeta.sharePath) {
-    throw new Error(`sharePath missing in pageMeta of ${pagePath}`);
-  }
-
-  return { route: pageMeta.sharePath, pageMeta };
-};
 
 const upsertMeta = (html, key, value, content) => {
   const escapedContent = content.replaceAll('"', "&quot;");
@@ -146,13 +79,12 @@ const prerenderRootMarkup = (pageMeta) => {
 };
 
 const run = async () => {
-  const pageModules = await findPageModules(path.resolve(process.cwd(), "src/features"));
+  const pageConfigs = await collectPageConfigs();
   const distDir = path.resolve(process.cwd(), "dist");
   const indexPath = path.join(distDir, "index.html");
   const template = await readFile(indexPath, "utf8");
 
-  for (const pagePath of pageModules) {
-    const { route, pageMeta } = await extractPageConfig(pagePath);
+  for (const { route, pageMeta } of pageConfigs) {
     const routeTemplate = withMeta(template, route, pageMeta).replace(
       '<div id="root"></div>',
       prerenderRootMarkup(pageMeta),
@@ -163,7 +95,7 @@ const run = async () => {
     await writeFile(path.join(outputDir, "index.html"), routeTemplate, "utf8");
   }
 
-  console.info(`Prerender completed for ${pageModules.length} routes.`);
+  console.info(`Prerender completed for ${pageConfigs.length} routes.`);
 };
 
 void run();
